libcompat: move deprecated funs out of API
Create a new internal library for deprecated functions, and initially
move the following functions there:
  flux_sendmsg()
  flux_recvmsg()
  flux_recvmsg_match()
These functions are no longer exported via the public API.
Include (temporaily) compat.h from in-tree <flux/core.h> include
file to avoid having to touch each user.

Update to flux_send, flux_recv the following that don't include
<flux/core.h>:
  t/loop/rpc.c
  t/loop/multrpc.c
  src/common/libjsonc/rpc.c
  src/common/libjsonc/request.c
