+++ configure.ac
@@ -127,6 +127,7 @@ AC_CONFIG_FILES( \
   src/common/libev/Makefile \
   src/common/libflux/Makefile \
   src/common/libjsonc/Makefile \
+  src/common/libcompat/Makefile \
   src/lib/Makefile \
   src/lib/libpmi/Makefile \
   src/bindings/Makefile \
